refactor: mastodon extract & rename function
