+++ lib/routes/mastodon/timeline.js
@@ -13,6 +13,6 @@ module.exports = async (ctx) => {
     ctx.state.data = {
         title: `Local Public${ctx.params.only_media ? ' Media' : ''} Timeline on ${site}`,
         link: `http://${site}`,
-        item: utils.ProcessFeed(list),
+        item: utils.parseStatuses(list),
     };
 };
