+++ lib/routes/mastodon/utils.js
@@ -1,4 +1,6 @@
-const ProcessFeed = (data) =>
+const got = require('@/utils/got');
+
+const parseStatuses = (data) =>
     data.map((item) => {
         const content = item.content ? item.content.replace(/<span.*?>|<\/span.*?>/gm, '') : '';
 
@@ -42,6 +44,82 @@ const ProcessFeed = (data) =>
         };
     });
 
+async function getAccountStatuses(site, account_id, only_media) {
+    const statuses_url = `http://${site}/api/v1/accounts/${account_id}/statuses?only_media=${only_media}`;
+    const statuses_response = await got({
+        method: 'get',
+        url: statuses_url,
+    });
+    let data = statuses_response.data;
+
+    let account_data;
+    if (data.length !== 0 && data[0].account !== null) {
+        account_data = data[0].account;
+    } else {
+        const account_url = `http://${site}/api/v1/accounts/${account_id}`;
+        const account_response = await got({
+            method: 'get',
+            url: account_url,
+        });
+        account_data = account_response.data;
+    }
+
+    if (only_media === true) {
+        data = data.filter((item) => item.media_attachments.length !== 0);
+    }
+    return { account_data, data };
+}
+
+async function getAccountIdByAcct(acct, ctx) {
+    const config = require('@/config').value;
+    const mastodonConfig = config.mastodon;
+
+    if (!(mastodonConfig.apiHost && mastodonConfig.accessToken && mastodonConfig.acctDomain)) {
+        throw 'this route require api configuration, please check <a href="https://docs.rsshub.app/en/install/#configuration">documentation</a>';
+    }
+
+    const site = mastodonConfig.apiHost;
+
+    const search_url = `http://${site}/api/v2/search`;
+    const cacheUid = `mastodon_acct_id/${site}/${acct}`;
+
+    const account_id = await ctx.cache.tryGet(cacheUid, async () => {
+        const search_response = await got({
+            method: 'get',
+            url: search_url,
+            headers: {
+                Authorization: `Bearer ${mastodonConfig.accessToken}`,
+            },
+            searchParams: {
+                q: acct,
+                type: 'accounts',
+            },
+        });
+        const [acctUser, acctHost] = acct.split('@').filter((e) => e);
+        let acctOnServer;
+
+        if (acctHost) {
+            if (acctHost === mastodonConfig.acctDomain) {
+                acctOnServer = acctUser;
+            } else {
+                acctOnServer = acctUser + '@' + acctHost;
+            }
+        } else {
+            acctOnServer = acctUser;
+        }
+
+        const accountData = search_response.data.accounts.filter((item) => item.acct === acctOnServer);
+
+        if (accountData.length === 0) {
+            throw `acct ${acct} not found`;
+        }
+        return accountData[0].id;
+    });
+    return { site, account_id };
+}
+
 module.exports = {
-    ProcessFeed: ProcessFeed,
+    parseStatuses: parseStatuses,
+    getAccountStatuses: getAccountStatuses,
+    getAccountIdByAcct: getAccountIdByAcct,
 };
